Automerge on soft conflict (#448)
* Move convert cipher command to tools command.

Closes #447.

* Make tests fail by detecting conflicts.

* Prepare enum for types of event logs.

* Prepare encodings for commit scan types.

* Prepare server handlers to scan event log commits.

* Prepare client scan() function.

* Initial test spec for scanning commits.

* Improve commit scan test specs.

* Return list of commit proofs in scan.

* Remove proof_at() function.

Prefer to manually create the proof in first_commit().

* Do not use Range for proof indices.

So we can support sparse indices.

* Initial logic to scan server proofs and find ancestor commit.

* Prepare endpoint for fetching event patch.

* Prepare client function to load event diff.

* Prepare to merge local and remote patches.

* Merge event records sorting by timestamp.

* Prepare rewind_local() and push_remote().

* Prepare server handler to patch event logs.

* Implement rewind_to() and add test spec.

* Prune obsolete commented types.

Experimental MultiTree and CommitRelationship types are not needed.

* Rename rewind_to() -> rewind().

* The after field of Diff is optional.

As when we will apply a merge after rewinding an event log we only have
Vec<T> and the proof and last commit hash, we don't know what the after
proof will be.

* Draft logic for rewind_local().

* Hide recovery module from docs as it's still work in progress.

* Acquire file lock in rewind().

* Improve server patch handler.

* Return CheckedPatch from merge operations.

* Add encoding for CheckedPatch.

* Initial server and client logic for patch requests.

* Improve CommitTree::proof_at() and use when scanning.

An auto merge remote push is now completing as expected, assertions
required in the test spec.

* Fix scan_commits test spec.

* Initial auto merge test spec.

* Improve assertions in test spec.

* Trigger HardConflict error.

* Remove before_change() in NetworkAccount.

* Assert on search index in test spec.

* Send websocket notifications in patch handler.

* Notify websockets when a patch is applied server-side.

* Remove patch_devices(), use standard sync().

* Add test spec for edit secrets auto merge.

* Add failing test spec for deleting secrets.

* Expose apply_records() on event logs.

So that we will be able to preserve time information during an auto
merge.

* Start refactor so that Patch stores EventRecord.

* Use EventRecord in Patch to preserve time information.

This is now critical for the auto merge logic to work as expected.

* Server rolls back rewind on merge conflict.

* Client rolls back rewind on merge conflict.

* Do not use proof_at() rebuild checkpoint proof on the client.

This fixes the conflict in the delete secrets auto merge test.

* Remove obsolete non-deterministic proof_at().

* Tidy type handling for commit tree hash.

* Remove unused serde implementation for CommitProof.

* Prune unused serde implementations.

* Update openapi docs.

* Add short circuit for hard conflict.

When the serve returns a proof for the first entry in an event log and
the client cannot verify the proof then we know the trees have diverged
so no need to scan all the proofs.

* Prepare to resolve hard conflict.

* Initial work on the force pull logic.

* Move hard conflict logic to auto merge module.

* Work in progress on hard conflict handling.

* Merge non-conflicting events on conflict.

Fixes the change_cipher and change_folder_password test specs.

* Merge non-conflicting folders on soft conflict.

* Remove unused after field.

* Remove commits from CheckedPatch::Success variant.

* Don't return commits in patch_unchecked().

* Do not return commit hashes in event log.

* Rename field in Diff.

* Using macros for auto merge impls.

* Improve auto merge/hard conflict macros.

* Remove unused ascending option.

We never scan from the beginning of an event log so keep things simple.

* Stub ForceMerge impl on ServerStorage.

* Draft implementation of ForceMerge for ServerStorage.

* Support all event log types in update_account().

So that the server-side implementation is consistent even if we are not
force pushing account, device and files event logs at the moment.

Later, we may wish to support compaction for these event logs.

* Comment on last_commit.

* Use diff_unchecked() in destructive changes.

* Refactor to use diff_checked().

So we have more formal implementation for how checkpoints are set.

* Rename diff_records() -> patch_records().

* Revert function name change.

* Rename to diff_events(), simplify diff_unchecked().

* Harden patch_replace() implementation.

So that the computed root hash is compared to the expected checkpoint
and if verification fails an attempt is made to rollback to a snapshot
of the event log.

* Restore test specs.

+++ crates/sdk/src/storage/server.rs
@@ -322,7 +322,9 @@ impl ServerStorage {
         #[cfg(feature = "files")]
         {
             let files_folder = self.paths.files_dir().join(id.to_string());
-            vfs::remove_dir_all(&files_folder).await?;
+            if vfs::try_exists(&files_folder).await? {
+                vfs::remove_dir_all(&files_folder).await?;
+            }
         }
 
         #[cfg(feature = "audit")]
@@ -390,9 +392,12 @@ impl ServerStorage {
     }
 
     /// Patch the devices event log.
+    #[deprecated]
     pub async fn patch_devices(&mut self, diff: &DeviceDiff) -> Result<()> {
         let mut event_log = self.device_log.write().await;
-        event_log.patch_checked(&diff.before, &diff.patch).await?;
+        event_log
+            .patch_checked(&diff.checkpoint, &diff.patch)
+            .await?;
 
         // Update in-memory cache of trusted devices
         let reducer = DeviceReducer::new(&event_log);
